# internal/server/gen/swagger.yaml
components:
  schemas:
...
    github_com_kevin_vigil_internal_httputil.ErrorResponse:
      properties:
        code:
          description: Code 机器可读错误码（可选，非 4xx 通用错误可留空）。
          example: invalid_argument
          type: string
        details:
          description: Details 结构化补充信息（可选，如字段级校验错误）。
        error:
          example: invalid id
          type: string
      type: object
...
    internal_ai.DiagnoseResult:
      properties:
        confidence:
          description: 置信度
          type: number
        evidence:
          description: 依据
          items:
            additionalProperties: {}
            type: object
          type: array
        insightID:
          description: AIInsight ID
          type: integer
        rootCause:
          description: 根因线索文本
          type: string
      type: object
    internal_ai.resolveReq:
      properties:
        accepted:
          type: boolean
      type: object
    internal_analytics.AlertMetrics:
      properties:
        noiseRate:
          description: 降噪率 = 1 - Notified/Total（0~1）
          type: number
        notified:
          description: 触发通知的（非噪音）
          type: integer
        total:
          description: 接入总量
          type: integer
        unrouted:
          description: 未命中路由
          type: integer
      type: object
    internal_analytics.Dashboard:
      properties:
        alert:
          $ref: '#/components/schemas/internal_analytics.AlertMetrics'
        incident:
          $ref: '#/components/schemas/internal_analytics.IncidentMetrics'
        load:
          items:
            $ref: '#/components/schemas/internal_analytics.TeamLoad'
          type: array
        postmortem:
          $ref: '#/components/schemas/internal_analytics.PostmortemMetrics'
      type: object
    internal_analytics.IncidentMetrics:
      properties:
        bySeverity:
          additionalProperties:
            type: integer
          description: critical/warning/info 各数量
          type: object
        byStatus:
          additionalProperties:
            type: integer
          type: object
        mttaratio:
          description: 平均确认时长（秒），无数据为 0
          type: number
        mttratio:
          description: 平均解决时长（秒）
          type: number
        resolvedCount:
          description: 已解决数（用于 MTTR 计算）
          type: integer
        total:
          type: integer
      type: object
    internal_analytics.PostmortemMetrics:
      properties:
        completionRate:
          description: published/total
          type: number
        published:
          type: integer
        total:
          type: integer
      type: object
    internal_analytics.TeamLoad:
      properties:
        incidents:
          description: 该团队事件数
          type: integer
        teamID:
          type: integer
        teamName:
          type: string
      type: object
    internal_analytics.TrendPoint:
      properties:
        date:
          description: YYYY-MM-DD
          type: string
        events:
          type: integer
        incidents:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 业务 API 推荐（唯一声明）方案：HTTP Bearer JWT（/auth/login 换取）。
      scheme: bearer
      type: http
...
info:
  description: |-
    Vigil 告警处置平台 REST API。
    认证：业务 API 要求 Bearer JWT（/auth/login 换取）。鉴权中间件另接受 X-Vigil-User-ID 头作为本地/回退身份（可伪造，仅限受信网络，生产禁用），该回退方案不在 securitySchemes 中声明。
  license:
    name: MIT
  title: Vigil API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /ai-insights/{id}/resolve:
    post:
      description: 人确认/拒绝 AI 建议（human-in-the-loop）。
      parameters:
      - description: AI Insight ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_ai.resolveReq'
                description: accepted=true 接受，false 拒绝
                summary: request
        description: accepted=true 接受，false 拒绝
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Resolve AI insight
      tags:
      - ai
  /analytics/alerts:
    get:
      description: 按 start/end（RFC3339）时间范围返回告警度量。
      parameters:
      - description: 起始时间 RFC3339
        in: query
        name: start
        schema:
          type: string
      - description: 结束时间 RFC3339
        in: query
        name: end
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_analytics.AlertMetrics'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get alert metrics
      tags:
      - analytics
  /analytics/dashboard:
    get:
      description: 返回近 N 天的仪表盘汇总（默认 N=7）。
      parameters:
      - description: 时间窗（天）
        in: query
        name: days
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_analytics.Dashboard'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get analytics dashboard
      tags:
      - analytics
  /analytics/incidents:
    get:
      description: 按 start/end（RFC3339）时间范围返回事件度量。
      parameters:
      - description: 起始时间 RFC3339
        in: query
        name: start
        schema:
          type: string
      - description: 结束时间 RFC3339
        in: query
        name: end
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_analytics.IncidentMetrics'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get incident metrics
      tags:
      - analytics
  /analytics/postmortems:
    get:
      description: 按 start/end（RFC3339）时间范围返回复盘度量。
      parameters:
      - description: 起始时间 RFC3339
        in: query
        name: start
        schema:
          type: string
      - description: 结束时间 RFC3339
        in: query
        name: end
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_analytics.PostmortemMetrics'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get postmortem metrics
      tags:
      - analytics
  /analytics/team-load:
    get:
      description: 按 start/end（RFC3339）时间范围返回各团队负载度量。
      parameters:
      - description: 起始时间 RFC3339
        in: query
        name: start
        schema:
          type: string
      - description: 结束时间 RFC3339
        in: query
        name: end
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/internal_analytics.TeamLoad'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get team load
      tags:
      - analytics
  /analytics/trend:
    get:
      description: 按 days/时间范围返回趋势点序列。
      parameters:
      - description: 时间窗（天）
        in: query
        name: days
        schema:
          type: integer
      - description: 起始时间 RFC3339
        in: query
        name: start
        schema:
          type: string
      - description: 结束时间 RFC3339
        in: query
        name: end
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/internal_analytics.TrendPoint'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get analytics trend
      tags:
      - analytics
...
  /incidents/{id}/diagnose:
    post:
      description: 触发 LLM 根因诊断；若未启用 LLM 返回 200 与 disabled 状态。
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: AI 诊断未启用
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_ai.DiagnoseResult'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Diagnose incident (AI)
      tags:
      - ai
...
  /incidents/{id}/similar:
    get:
      description: 按向量/文本相似度查询与给定 incident 相似的历史事件。
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: 返回条数上限
        in: query
        name: limit
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: '{similar: []*ent.Incident}'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Find similar incidents
      tags:
      - ai
...
servers:
- url: /api/v1